Conflicts:
	trunk/ServerCore/Stump.Server.WorldServer/Handlers/Character/CharacterCreationHandler.cs
	trunk/ServerCore/Stump.Server.WorldServer/Handlers/Character/CharacterDeletionHandler.cs
	trunk/ServerCore/Stump.Server.WorldServer/Handlers/Character/CharacterStatsHandler.cs

+++ trunk/ServerCore/Stump.Server.WorldServer/Handlers/Inventory/InventorySpellsHandler.cs
@@ -24,7 +24,7 @@ namespace Stump.Server.WorldServer.Handlers
 {
     public partial class InventoryHandler : WorldHandlerContainer
     {
-        [WorldHandler(typeof(SpellMoveMessage))]
+        [WorldHandler(typeof (SpellMoveMessage))]
         public static void HandleSpellMoveMessage(WorldClient client, SpellMoveMessage message)
         {
             if (message.position < 63 || message.position > 255)
@@ -32,14 +32,14 @@ namespace Stump.Server.WorldServer.Handlers
                 return;
             }
 
-            client.ActiveCharacter.ModifySpellPos((SpellIdEnum)message.spellId, (int) message.position);
+            client.ActiveCharacter.ModifySpellPos((SpellIdEnum) message.spellId, (int) message.position);
 
             SendSpellMovementMessage(client, message.spellId, message.position);
         }
 
         public static void SendSpellMovementMessage(WorldClient client, uint spellId, uint position)
         {
-            client.Send(new SpellMovementMessage((uint) spellId, position));
+            client.Send(new SpellMovementMessage(spellId, position));
         }
 
         public static void SendSpellListMessage(WorldClient client, bool previsualization)
